Add programs for tests
There are 3 tipes of Fibonacci tests. First "Fibonacci.c" is a 20
number sequence within a vector.
Second "Fibonacci2.c" is the same but onely with registers.
Third "Fibonacci3.c" is uses recursive method to calculate.
"mult.c" is a multiplication in a 'for' loop.
The Makefile it depends of 'mult.c' for now.