Change structure of repo
The Adept processor will be tested with a variety of programs. To ease
the process of testing and to expedite the remote testing framework, all
files in this repo are generated and tested against the core every time
a PR is made.

Binaries are stores in bin/, elfs in elf/, and hexs in hex/. Moved all
simple programs to their corresponding directory.